String manipulation completed

+++ 0x0B-malloc_free/_strlen.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-size_t _strlen(const char *s) {
+int _strlen(const char *s) {
     size_t length = 0;
     while (*s != '\0') {
         length++;
